refactor: move PayoutRow to its own file
