+++ src/components/PayoutList/index.tsx
@@ -1,36 +1,17 @@
 import React, { useMemo, useState, useEffect } from "react";
-import { formatUnits } from "ethers/utils";
 
-import { TableRow, TableCell, TableBody } from "@material-ui/core";
-
+import { TableBody } from "@material-ui/core";
 import { ColonyRole } from "@colony/colony-js";
 
+import PayoutRow from "./PayoutRow";
 import NewPayoutRow from "./NewPayoutRow";
 import Table from "../common/StyledTable";
-import PayoutModal from "../Modals/PayoutModal";
 
 import { useSafeInfo } from "../../contexts/SafeContext";
-import { useColonyClient, useHasDomainPermission, useTokens } from "../../contexts/ColonyContext";
+import { useColonyClient, useHasDomainPermission } from "../../contexts/ColonyContext";
 
 import getActivePayouts from "../../utils/colony/getColonyPayouts";
 import { PayoutInfo } from "../../typings";
-
-const PayoutRow = ({ payout }: { payout: PayoutInfo }) => {
-  const tokens = useTokens();
-  const payoutToken = tokens.find(token => token.address === payout.tokenAddress);
-  const [isOpen, setIsOpen] = useState<boolean>(false);
-
-  if (!payoutToken) return null;
-  return (
-    <>
-      <PayoutModal isOpen={isOpen} setIsOpen={setIsOpen} payout={payout} token={payoutToken} />
-      <TableRow onClick={() => setIsOpen(true)}>
-        <TableCell>{payoutToken?.symbol || payout.tokenAddress}</TableCell>
-        <TableCell align="right">{formatUnits(payout.amount, payoutToken?.decimals)}</TableCell>
-      </TableRow>
-    </>
-  );
-};
 
 const PayoutList = () => {
   const safeInfo = useSafeInfo();
